+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Version History
 
+## 0.5.6_preview / 2022-01-13
+
+- Implemented additional_headers parameter to base client's request function
+- Added verbosity and community id header wrapper functions
+
 ## 0.5.5_preview / 2022-01-14
 
 - Reinstated support for Python 3.7